requirements.txt
pydantic>=1.8.0,<2.0.0
fastapi>=0.110.0
uvicorn>=0.28.0
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
aiokafka>=0.10.0
pymongo>=4.6.2
kafka-python>=2.0.2

# Opentelemetry instrumentation
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-asyncio==0.44b0
opentelemetry-instrumentation-aws-lambda==0.44b0
opentelemetry-instrumentation-confluent-kafka==0.44b0
opentelemetry-instrumentation-dbapi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-grpc==0.44b0
opentelemetry-instrumentation-kafka-python==0.44b0
opentelemetry-instrumentation-logging==0.44b0
opentelemetry-instrumentation-pymongo==0.44b0
opentelemetry-instrumentation-sqlalchemy==0.44b0
opentelemetry-instrumentation-sqlite3==0.44b0
opentelemetry-instrumentation-system-metrics==0.44b0
opentelemetry-instrumentation-tortoiseorm==0.44b0
opentelemetry-instrumentation-urllib==0.44b0
opentelemetry-instrumentation-wsgi==0.44b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-propagator-b3==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-test-utils==0.44b0
opentelemetry-util-http==0.44b0
splunk-opentelemetry==1.18.0